Create startRound function
This function takes 2 arguments, the players choice and the computers
choice. It then compares the two to see if they are equal, if it is,
the round is a tie. If they are not equal, then they are both compared
to the rules of rps to decide who wins the round. The function will
return string with who won the round.