Create playGame function
This function starts the rps game with 5 rounds. It also keeps track of
the computer and user scores to see who won the match out of the 5
rounds, incrementing each winner by 1 point. Once the 5 rounds are over,
whoever has the most points will be the winner, and a message will print
to the console of who won. If the game ends in a tie, a message will
print to the console stating that the game has ended in a tie.