Create function to decide who won 5 rounds of RPS
the decideWinner function tests to see if the user or computer has
reached a score of 5. If so, two display messages will show. One
announceing if the user won or lost, and one congratulating or telling
the user they lost, and also showing the final score. This function also
clears the scores back to 0, along with the display of the scores
reverting back to 0;